commit 12

+++ app/Http/Controllers/TodoController.php
@@ -12,13 +12,13 @@ class TodoController extends Controller
     {
         $today = date('Y-m-d');
         $datas = Todo::whereUserId(auth()->user()->id)->where('date', $today)->get();
-
+        
         return view('todo.index', [
             'title' => 'Todo List',
             'datas' => $datas,
         ]);
     }
-
+    
     public function store(Request $request)
     {
         $validatedData = $request->validate([
@@ -28,28 +28,36 @@ class TodoController extends Controller
         $validatedData['slug'] = substr(base_convert(sha1(uniqid(mt_rand())), 16, 36), 0, 10);
         $validatedData['content'] = Crypt::encryptString($request->content);
         $validatedData['date'] = date('Y-m-d');
-
+        
         Todo::create($validatedData);
         return back();
     }
-
+    
     public function update(Todo $todo)
     {
-        $todo->update(['is_done' => true]);
-        return back();
+        if($todo->user_id == auth()->user()->id) {
+            $todo->update(['is_done' => true]);
+            return back();
+        } else {
+            return to_route('todo.index');
+        }
     }
-
+    
     public function destroy(Todo $todo)
     {
-        $todo->delete();
-        return back();
+        if($todo->user_id == auth()->user()->id) {
+            $todo->delete();
+            return back();
+        } else {
+            return to_route('todo.index');
+        }
     }
-
+    
     public function history()
     {
         $today = date('Y-m-d');
         $datas = Todo::whereUserId(auth()->user()->id)->where('date', '!=', $today)->latest()->paginate(20);
-
+        
         return view('todo.history', [
             'title' => 'History List',
             'datas' => $datas,
